Use regex in isUrlAbsolute
This function will return true if the url starts with either "http://" or "https://", and false otherwise. The regular expression ^https?:\/\/ matches strings that start with "http://" or "https://". The caret (^) specifies the start of the string, and the question mark (?) makes the "s" optional. The double forward slashes are escaped with backslashes, as forward slash is a special character in regular expressions