added DynamicEncounter generation
The update aims to give the ability to generate an encounter dynamically by passing the parameters. This allows to generate encounters by using macro or other modules. To achieve this the generateDynamicEncounter function has been created and exposed. It replicates what the dialog window does. the createDynamicEncounters function generates 1 encounter with the parameters that are passed to it. Finally the DynamicSpawn function generates the spawning given a template as input